requirements.txt
flask
waitress
transformers
torch
python-dotenv